+++ code/branches/presentation2/src/orxonox/items/MultiStateEngine.h
@@ -70,7 +70,7 @@ namespace orxonox
             const std::string& getDefEngSndBoost();
 
         private:
-            int steeringDirectionZ_;
+            float steeringDirectionZ_;
             int state_;
             LuaState* lua_;
             std::vector<EffectContainer*> effectContainers_;
